% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apiUrl}
\alias{apiUrl}
\title{Get gnomAD API URL}
\usage{
apiUrl()
}
\value{
Character with gnomAD API URL
}
\description{
Returns the gnomAD API URL as a character. Checks if the option
'gnomadr.apiurl' has been set to allow for hypothetical parameterization
in the future
}
