% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVariantIds.R
\name{convertVariantIds}
\alias{convertVariantIds}
\title{Get variant IDs for alternate reference genome}
\usage{
convertVariantIds(varids, genomes)
}
\arguments{
\item{varids}{Character, gnomAD variant_id(s)}

\item{genomes}{The genome associated with the given variant_id}
}
\value{
variant_id(s)
}
\description{
Returns the variant_id(s) for the alternate reference genome
}
\examples{
\dontrun{
convertVariantIds(varids = '1-89388944-A-C', genomes = "GRCh37")
}

}
