% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantPopData.R
\name{getVariantPopData}
\alias{getVariantPopData}
\title{Get variant population data}
\usage{
getVariantPopData(varids, genomes)
}
\arguments{
\item{varids}{Character, gnomAD variant_id(s)}

\item{genomes}{The genome(s) associated with the given variant_id}
}
\value{
List of population data for the given variant_id(s)
}
\description{
Returns the variant_id(s) for the alternate reference genome
}
\examples{
\dontrun{
getVariantPopData(varids = '1-88923261-A-C', genomes = "GRCh38")
}

}
