% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomadR-package.R
\docType{package}
\name{gnomadR-package}
\alias{gnomadR}
\alias{gnomadR-package}
\title{gnomadR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daynefiler/gnomadR}
  \item Report bugs at \url{https://github.com/daynefiler/gnomadR/issues}
}

}
\author{
\strong{Maintainer}: Dayne L Filer \email{dayne.filer@gmail.com} (\href{https://orcid.org/0000-0002-3443-5315}{ORCID})

}
\keyword{internal}
