% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomadR-package.R
\name{pkgParams}
\alias{pkgParams}
\title{Common function parameters}
\arguments{
\item{dataset}{Character scalar, gnomAD DatasetId}

\item{datasets}{Character, gnomAD DatasetId(s)}

\item{genome}{Character scalar, gnomAD ReferenceGenomeId}

\item{genomes}{Character, gnomAD ReferenceGenomeId(s)}

\item{chroms}{Character, the chromosome(s) ('1', '2', ..., 'X', 'Y')}

\item{chrom}{Character scalar, the chromosome ('1', '2', ..., 'X', 'Y')}

\item{starts}{Integer, starting position(s)}

\item{start}{Integer scalar, starting position}

\item{stop}{Integer scalar, ending position}

\item{varid}{Charactter scalar, gnomAD variant_id}

\item{varids}{Character, gnomAD variant_id(s)}
}
\description{
Common function parameters
}
